# Bump Wazuh dashboard to 2.8.0