--- a/README.md
+++ b/README.md
@@ -36,7 +36,8 @@ The following table shows the references for the versions of each component.
 |-----------------|-----------------------|
 | 4.3.x           | 1.2.0                 |
 | 4.4.0           | 2.4.1                 |
-| 4.4.1 - 4.5.1   | 2.6.0                 |
+| 4.4.1 - 4.5.2   | 2.6.0                 |
+| 4.6.0           | 2.8.0                 |
 
 ### Wazuh indexer