--- a/stack/dashboard/base/files/etc/services/wazuh-dashboard.service
+++ b/stack/dashboard/base/files/etc/services/wazuh-dashboard.service
@@ -7,7 +7,7 @@ User=wazuh-dashboard
 Group=wazuh-dashboard
 EnvironmentFile=-/etc/default/wazuh-dashboard
 EnvironmentFile=-/etc/sysconfig/wazuh-dashboard
-ExecStart=/usr/share/wazuh-dashboard/bin/opensearch-dashboards "-c /etc/wazuh-dashboard/opensearch_dashboards.yml"
+ExecStart=/usr/share/wazuh-dashboard/bin/opensearch-dashboards
 WorkingDirectory=/usr/share/wazuh-dashboard
 
 [Install]