--- a/stack/dashboard/rpm/docker/x86_64/Dockerfile
+++ b/stack/dashboard/rpm/docker/x86_64/Dockerfile
@@ -1,13 +1,16 @@
-FROM rockylinux:8.5
+FROM amd64/centos:7
+
+# Enable EPEL
+RUN yum install -y http://packages.wazuh.com/utils/pkg/epel-release-latest-7.noarch.rpm
 
 # Install all the necessary tools to build the packages
 RUN yum clean all && yum update -y
 RUN yum install -y openssh-clients sudo gnupg \
-    yum-utils epel-release redhat-rpm-config rpm-devel \
+    yum-utils redhat-rpm-config rpm-devel \
     zlib zlib-devel rpm-build autoconf automake \
     glibc-devel libtool perl
 
-RUN yum install -y https://repo.ius.io/ius-release-el$(rpm -E '%{rhel}').rpm  
+RUN yum install -y https://repo.ius.io/ius-release-el$(rpm -E '%{rhel}').rpm
 
 RUN yum update -y && yum install -y python3